Created more stemming tools for workflow api
Used nltk stemmers such as Porter, Snowball, and Lancaster stemmer. Files to test with are included as well as their output.
Right now the proper workflow would be to:
Clean -> Tokenize -> Stem -> Classify the tweets in that order

Next plan of action for new tools would be to include additional tokenizing tools with different methods.